contar usos
Step 1. Create and complete the implementation file "pedir-nombre-fichero.cpp" corresponding to the interface file "pedir-nombre-fichero.hpp". The objective of the function pedirNombreFichero, declared in the interface file, is to assist in asking the user for the name of a file (such as "usos-16.csv" or "pruebas-2000.csv") and convert it into a relative path to the project's execution directory as required in this task and the next one.
If you are using Visual Studio Code, you should not need to modify the value of the constant RUTA_RELATIVA declared in the interface file. Otherwise, you can modify it if necessary.

Step 2. The file "uso-bizi.hpp" defines the interface of a module for working with records representing usages of the Bizi Zaragoza system. Complete the definition of the UsoBizi record type so that it represents the following data from a usage of the Bizi Zaragoza system:

the identifier of the user who uses the bicycle;
the code of the station where the bicycle is taken; and
the code of the station where the bicycle is returned.
Note: The definition of this record type is not strictly necessary for solving this problem, although it simplifies file reading operations in the programs required in this task and the next.

Step 3. Create and complete the implementation file "uso-bizi.cpp" corresponding to the interface file "uso-bizi.hpp".

Step 4. Write the main module of the project with its main() function in the file "2-contar-usos.cpp".